Fixed broken link as reported in issue #35
Fixed broken link to the list of supported PS/GS fonts with the Painters renderer. Apparently MathWorks modified the documentation in R2014b and this list was removed. It is unclear at this point whether this is because they have removed the fonts limitation in 14b, or simply because they have decided to remove this documentation section.