BUG: Fix for MATLAB lib files older than pdftops
This issue was found running MATLAB R2014a on Ubuntu 15.10.
The error message was:

    Error running pdftops:
    pdftops: /usr/local/MATLAB/R2014a/bin/glnxa64/libtiff.so.5: no version information available (required by /usr/lib/x86_64-linux-gnu/libpoppler.so.52)
    pdftops: /usr/local/MATLAB/R2014a/sys/os/glnxa64/libstdc++.so.6: version `GLIBCXX_3.4.20' not found (required by /usr/lib/x86_64-linux-gnu/libpoppler.so.52)

And it occurs because the libtiff and libstdc++ files in MATLAB
R2014a's bin are older than those expected the 2015 version of
pdftops.

The fix is the same as implemented with the subfunction `gs_command`
in ghostscript.m (added in 6c11f9f). Namely we clear the
`LD_LIBRARY_PATH` environment variable before calling the shell
command (pdftops) to ensure it is looking at the system lib files
instead of the MATLAB lib files.